Add animals tag, separate categories, and bug fixs
Animals category was added to the CATEGORIES array
by special request, additionally I removed the
array from the App.js file and put it into a
separate file.

I'll probably move that to a fixtures or static
data folder once I remove the rest of the
components from App.js and place them into
other folders.

Bug Fix:
Removed a reference to manifest.json in the
index.html file, it was causing issues since
the file does not exist.