Add unit tests for Category and CategoryFilter
While creating tests for CategoryFilter I also
turned the code for the categories into a
component (named Category). The code looks far
cleaner as a result.

There are some unit tests for both Category
and Category filter in this commit, and
I'm going to add more for them and the other
components.

Small update to .vscode with launch.json for
using the built in debugger for jest components.

Update to package.json to update the run test
command to look for leaks.